# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Periodico, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Publicaciones/ListarPorCategoria/{idCategoria}": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idCategoria",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "estado",
            "type": "string",
            "default": "Activo"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Publicacion"
              }
            }
          }
        }
      }
    },
...
    "/api/Publicaciones/{id}": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Publicacion"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Publicaciones/Actualizar": {
      "put": {
        "tags": [
          "Publicaciones"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Titulo",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Descripcion",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Fecha",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "idcategoria",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Imagen",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Publicaciones/ActualizarSImg": {
      "put": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "titulo",
            "type": "string"
          },
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          },
          {
            "in": "query",
            "name": "fecha",
            "type": "string"
          },
          {
            "in": "query",
            "name": "idcategoria",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "imagen",
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Publicaciones/Crear": {
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Titulo",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Descripcion",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Fecha",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "idusuario",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "idcategoria",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Imagen",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Rol"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Roles/Actualizar": {
      "put": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Roles/Crear": {
      "post": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Publicacion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titulo": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "visualizacion": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idcategoria": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Rol": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "usuariosRoles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UsuarioRol"
          }
        }
      }
    },
...
    "UsuarioRol": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idrol": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}